#67, Refatoração estrutural dos módulos de requisição de exames e páginas web do SISCAN
- Reorganização dos imports para refletir nova arquitetura de pacotes (src.siscan.requisicao.*, src.siscan.webpage.*, etc.).
- Modularização e centralização de métodos utilitários na classe abstrata WebPage (src/siscan/webpage/base.py).
- Ajuste das classes de requisição de exame para utilizar os novos utilitários e estrutura de herança.
- Atualização dos schemas Pydantic com novas regras condicionais de validação e restrições de tamanho de campos.
- Correção de validações de campos cirúrgicos e dependências radioterápicas no schema de mamografia.
- Atualização dos testes automatizados para a nova estrutura de módulos e imports.
- Remoção de código duplicado e melhorias gerais na organização do código.
