+++ src/siscan/requisicao/requisicao_exame.py
@@ -1,15 +1,13 @@
+import logging
+from abc import abstractmethod
+from pydantic import BaseModel
 from typing import Type, Any
 
-from pydantic import BaseModel
-
-from abc import abstractmethod
-
-import logging
-
 from src.siscan.exception import SiscanInvalidFieldValueError
-from src.siscan.classes.webpage import SiscanWebPage
+from src.siscan.webpage.base import SiscanWebPage
+from src.siscan.webpage.xpath_constructor import (XPathConstructor as XPE,
+                                                  InputType)
 from src.utils.SchemaMapExtractor import SchemaMapExtractor
-from src.utils.xpath_constructor import XPathConstructor as XPE, InputType
 
 logger = logging.getLogger(__name__)
