+++ tests/test_env_user_auth.py
@@ -5,10 +5,10 @@ import secrets
 from src.main import app
 from src.env import SISCAN_USER, SISCAN_PASSWORD, SISCAN_URL, get_db
 from src.models import User, ApiKey
-from src.siscan.classes.requisicao_exame_mamografia_rastreio import (
+from src.siscan.requisicao.requisicao_exame_mamografia_rastreio import (
     RequisicaoExameMamografiaRastreio,
 )
-from src.siscan.context import SiscanBrowserContext
+from src.siscan.webpage.context import SiscanBrowserContext
 
 
 @pytest.fixture(scope="module")
