+++ tests/test_map_schema_fields.py
@@ -2,7 +2,7 @@ import pytest
 
 
 def test_requiscao_mamografia_map_schema_fields():
-    from src.siscan.classes.requisicao_exame_mamografia import RequisicaoExameMamografia
+    from src.siscan.requisicao.requisicao_exame_mamografia import RequisicaoExameMamografia
 
     assert set(RequisicaoExameMamografia.MAP_SCHEMA_FIELDS) == {
         "num_prontuario",
@@ -48,7 +48,7 @@ def test_requiscao_mamografia_map_schema_fields():
     }
 
 def test_requiscao_mamografia_rastreamento_map_schema_fields():
-    from src.siscan.classes.requisicao_exame_mamografia_rastreio import RequisicaoExameMamografiaRastreio
+    from src.siscan.requisicao.requisicao_exame_mamografia_rastreio import RequisicaoExameMamografiaRastreio
 
     assert set(RequisicaoExameMamografiaRastreio.MAP_SCHEMA_FIELDS) == {
         "tipo_mamografia_de_rastreamento",
